requirements.txt
django
djangorestframework
environs
requests
drf-yasg